Enhancements - only allow to edit or delete by the owner of the image

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,23 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
+        User currentUser = (User) session.getAttribute("loggeduser");
+        User imgOwnerUser = image.getUser();
+        boolean authorizedUser = checkUserAuthorization(currentUser, imgOwnerUser);
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        if(authorizedUser) {
+           return "images/edit";
+        }
+        else {
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +150,26 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model,HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        User currentUser = (User) session.getAttribute("loggeduser");
+        User imgOwnerUser = image.getUser();
+        boolean authorizedUser = checkUserAuthorization(currentUser, imgOwnerUser);
+        if(authorizedUser) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            String error = "Only the owner of the image can delete the image";
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
+
+
+
     }
 
 
@@ -182,9 +209,23 @@ public class ImageController {
             tagString.append(tags.get(i).getName()).append(",");
         }
 
-        Tag lastTag = tags.get(tags.size() - 1);
-        tagString.append(lastTag.getName());
 
+        if(tags.size() == 0) {
+            tagString = tagString.append("");
+        }
+        else {
+            Tag lastTag = tags.get(tags.size() - 1);
+            tagString.append(lastTag.getName());
+        }
         return tagString.toString();
+    }
+
+    private boolean checkUserAuthorization(User currUser, User imgOwn) {
+        if(currUser.getId() == imgOwn.getId()) {
+            return true;
+        }
+        else {
+            return false;
+        }
     }
 }
